Add files via upload
The first part of my project aims to create some functions.
I used to use Python when I needed to code, and some packages were useful to be faster.
Here, I create a function which is the Cumulative Distribution Function of a Gaussian function : https://fr.wikipedia.org/wiki/Fonction_d%27erreur
I also create the function linspace, which is equivalent with python from the numpy's package. 

In terms of class, I create the class BS which need 6 arguments :
- The spot S,
- The strike K,
- The volatility sigma,
- The rate free-risk r,
- The maturity T,
- The divided q (or div).

Both d1 and d2 parameters are computed in the class, and are memorized by the structure "param". And the last method of the class allows to compute the call or put price, according to the parameter opt (which is a boolean).